[FEAT] 영화 조회 API 구현

+++ src/main/java/com/ceos/spring_cgv_23rd/domain/movie/entity/MovieStatistic.java
@@ -30,7 +30,7 @@ public class MovieStatistic {
 
     @Builder.Default
     @Column(name = "view_count", nullable = false)
-    private Integer viewCount = 0;
+    private Long viewCount = 0L;
 
     @Builder.Default
     @Column(name = "egg_count", nullable = false)
